# backend/aviatur-api/swagger.json
{"openapi":"3.0.0","info":{"title":"aviatur_api","description":"API REST para el registro de hoteles en la compañía AVIATUR.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"HotelDto":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"stars":{"type":"number"},"price":{"type":"number"},"image":{"type":"string"},"amenities":{"type":"array","items":{"type":"string"}}},"required":["_id","name","stars","price","image","amenities"]}}},"paths":{"/hoteles":{"post":{"operationId":"HotelesController_post","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HotelDto"}}}},"responses":{"201":{"description":""}},"tags":["hoteles"]},"get":{"operationId":"HotelesController_getAll","parameters":[{"name":"query","required":false,"in":"query","description":"query - Filter. e.g. col1:v1,col2:v2 …","schema":{"type":"string"}},{"name":"fields","required":false,"in":"query","description":"fields - Fields returned. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"sortby","required":false,"in":"query","description":"sortby - Sorted-by fields. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"order","required":false,"in":"query","description":"order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"limit - Limit the size of result set. Must be an integer","schema":{"type":"string"}},{"name":"offset","required":false,"in":"query","description":"offset - Start position of result set. Must be an integer","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["hoteles"]}},"/hoteles/{id}":{"get":{"operationId":"HotelesController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["hoteles"]},"put":{"operationId":"HotelesController_put","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HotelDto"}}}},"responses":{"200":{"description":""}},"tags":["hoteles"]},"delete":{"operationId":"HotelesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["hoteles"]}}}}